Got a bit more working
