+++ packages/next-intl/src/server/index.tsx
@@ -3,4 +3,4 @@
  */
 
 export {default as createIntlMiddleware} from './createIntlMiddleware';
-export {default as NextIntlServerProvider} from './NextIntlServerProvider';
+export {NextIntlServerProvider} from './NextIntlServerRuntimeContext';
